LayoutTests/js/string-split-double-empty-expected.txt

This test checks for a regression against split() function returns wrong answer for second empty split.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "".split(/s+/) is [""]
PASS "".split(/s+/) is [""]
PASS successfullyParsed is true

TEST COMPLETE
